requirements.txt
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
firebase-admin==5.0.3
Flask==2.0.1
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.3
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
httplib2==0.19.1
idna==3.2
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
msgpack==1.0.2
networkx==2.6.2
numpy==1.21.2
packaging==21.0
pandas==1.3.2
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
threadpoolctl==2.2.0
tifffile==2021.8.30
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.1
kmodes==0.11.1
